move ownership logic to specific module
